update products and categories api 2

+++ app/Containers/AppSection/Product/Tasks/DeleteProductTask.php
@@ -3,11 +3,13 @@
 namespace App\Containers\AppSection\Product\Tasks;
 
 use App\Containers\AppSection\Product\Data\Repositories\ProductRepository;
+use App\Containers\AppSection\Product\Models\Product;
 use App\Ship\Exceptions\DeleteResourceFailedException;
 use App\Ship\Exceptions\NotFoundException;
 use App\Ship\Parents\Tasks\Task as ParentTask;
 use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Http\JsonResponse;
 
 class DeleteProductTask extends ParentTask
 {
@@ -20,9 +22,17 @@ class DeleteProductTask extends ParentTask
      * @throws DeleteResourceFailedException
      * @throws NotFoundException
      */
-    public function run($id): int
+    public function run($id): JsonResponse | int
     {
         try {
+            $product = Product::query()->findOrFail($id);
+            if ($product == null) {
+                $response = [
+                    'message' => 'Product not found',
+                ];
+                return response()->json($response, 404);
+            }
+            $product->deleteImages();
             return $this->repository->delete($id);
         } catch (ModelNotFoundException) {
             throw new NotFoundException();
